+++ lib/trailblazer/activity/task_wrap/pipeline.rb
@@ -17,7 +17,7 @@ class Trailblazer::Activity
       attr_reader :sequence
 
       def self.insert_before(pipe, before_id, insertion)
-        index = pipe.sequence.find_index { |(id, _)| id == before_id }
+        index = find_index(pipe, before_id)
 
         seq = pipe.sequence.dup
 
@@ -25,16 +25,30 @@ class Trailblazer::Activity
       end
 
       def self.insert_after(pipe, after_id, insertion)
-        index = pipe.sequence.find_index { |(id, _)| id == after_id }
+        index = find_index(pipe, after_id)
 
         seq = pipe.sequence.dup
 
         Pipeline.new(seq.insert(index+1, insertion))
       end
 
-      def self.append(pipe, _, insertion) # TODO: test me.
+      def self.append(pipe, _, insertion)
         Pipeline.new(pipe.sequence + [insertion])
       end
+
+      def self.prepend(pipe, insertion_id, insertion, replace: 0)
+        return Pipeline.new(insertion.to_a + pipe.sequence) if insertion_id.nil?
+
+        index = find_index(pipe, insertion_id)
+
+        Pipeline.new(pipe.sequence[0..index-1] + insertion.to_a + pipe.sequence[index+replace..-1])
+      end
+
+      # @private
+      def self.find_index(pipe, id)
+        index = pipe.sequence.find_index { |(seq_id, _)| seq_id == id }
+      end
+
 
       # Merges {extension_rows} into the {Pipeline} instance.
       # This is usually used in step extensions or at runtime for {wrap_runtime}.